도서 도메인 모델 구조 변경 - root : BookInventory
